Instalando SweetAlert y mostrando alerta al crear, editar y elimnar cliente

+++ src/components/Clientes.jsx
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom'
 import { Query, Mutation } from 'react-apollo';
 import { CLIENTES_QUERY } from '../queries';
 import { ELIMINAR_CLIENTE } from '../mutations';
+import Swal from 'sweetalert2'
 
 const Clientes = () => (
   <Query query={CLIENTES_QUERY} pollInterval={1000}>
@@ -29,8 +30,26 @@ const Clientes = () => (
                           type="button"
                           className="btn btn-danger db-block d-md-inline-block mr-2"
                           onClick={() => {
-                            eliminarCliente({
-                              variables: { id }
+                            Swal.fire({
+                              title: '¿Estas seguro?',
+                              text: "¡No podrás revertir esto!",
+                              icon: 'warning',
+                              showCancelButton: true,
+                              confirmButtonColor: '#3085d6',
+                              cancelButtonColor: '#d33',
+                              cancelButtonText: 'Cancelar',
+                              confirmButtonText: '¡Si, eliminar!'
+                            }).then((result) => {
+                              if (result.value) {
+                                eliminarCliente({
+                                  variables: { id }
+                                })
+                                Swal.fire(
+                                  '¡Eliminado!',
+                                  'El cliente ha sido eliminado.',
+                                  'success'
+                                )
+                              }
                             })
                           }}
                           >
